+++ src/app/modules/candidate-registration/candidate-registration.module.ts
@@ -1,22 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
 import { CandidateRegistrationRoutingModule } from './candidate-registration-routing.module';
-import { RegistrationFormComponent } from './registration-form/registration-form.component';
+import { RegistrationFormComponent } from './components/registration-form/registration-form.component';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { NgMultiSelectDropDownModule } from 'ng-multiselect-dropdown';
 
-
 @NgModule({
-  declarations: [
-    RegistrationFormComponent
-  ],
+  declarations: [RegistrationFormComponent],
   imports: [
     CommonModule,
     FormsModule,
     ReactiveFormsModule,
     NgMultiSelectDropDownModule.forRoot(),
-    CandidateRegistrationRoutingModule
-  ]
+    CandidateRegistrationRoutingModule,
+  ],
 })
-export class CandidateRegistrationModule { }
+export class CandidateRegistrationModule {}
