+++ src/app/modules/candidate-registration/components/registration-form/registration-form.component.spec.ts
